requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.0
pillow==10.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.0
aiofiles==23.2.1
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
